1) Correction of lower case query for plastn (dust masker previously allowed only upper case) 2) Correction float => double for evalue  3) Correction of score matrix

+++ src/algo/core/api/IAlgoParameters.hpp
@@ -121,7 +121,7 @@ public:
     int     extendGapCost;
 
     /** Expected value to be used. */
-    float   evalue;
+    double   evalue;
 
     /** Ungapped X drop off. */
     int XdroppofUnGap;
